+++ src/Sql/Sql/ChangeLog.md
@@ -18,8 +18,11 @@
         - Additional information about change #1
 -->
 ## Upcoming Release
+
+## Version 2.3.0
 * Added support for cross subscription point in time restore on Managed Instances.
-* Add support for changing existing Sql Managed Instance hardware generation
+* Added support for changing existing Sql Managed Instance hardware generation
+* Fixed `Update-AzSqlServerVulnerabilityAssessmentSetting` help examples: parameter/property output - EmailAdmins
 
 ## Version 2.2.0
 Fix New-AzSqlDatabaseSecondary cmdlet to check for PartnerDatabaseName existence instead of DatabaseName existence.